#spring3hw5

+++ src/main/java/com/edu/ulab/app/dto/UserDto.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserDto {
-    private Long id;
+    private int id;
     private String fullName;
     private String title;
     private int age;
